+++ lib/l10n/intl_en.arb
@@ -1,0 +1,4 @@
+{
+  "@@locale" : "en",
+  "hello_world":"Hello world"
+}
