+++ lib/l10n/intl_vi.arb
@@ -1,0 +1,4 @@
+{
+  "@@locale" : "vi",
+  "hello_world":"Xin chào thế giới"
+}
